fix: 修復多項資料庫功能與回測依賴問題
## 修復內容

### 1. 知識庫分類追蹤功能修復
- 新增 `intent_classified_at` 和 `needs_reclassify` 欄位
- 建立 `database/init/13-add-knowledge-classification-tracking.sql`
- 建立 `database/fixes/add_knowledge_classification_tracking.sql`
- 修復意圖分配頁面統計載入失敗問題
- 新增 `scripts/verify_classification_tracking.py` 驗證工具

### 2. 知識匯入相似度檢查修復
- 新增 4 個相似度檢查函數：
  - find_similar_knowledge()
  - find_similar_knowledge_candidate()
  - find_similar_test_scenario()
  - check_knowledge_exists_by_similarity()
- 新增 test_scenarios.question_embedding 欄位
- 修復類型轉換問題（知識匯入服務）
- 新增 `scripts/generate_test_scenario_embeddings.py`
- 新增 `scripts/verify_similarity_functions.py`

### 3. AI 知識審核功能修復
- 更新 approve_ai_knowledge_candidate 為 4 參數版本
- 修正欄位名稱：linked_knowledge_ids → related_knowledge_ids
- 移除 has_knowledge 欄位引用
- 支援多意圖映射、embedding 複製、編輯版本選擇
- 更新 init 腳本確保未來一致性
- 新增 `scripts/comprehensive_approval_check.py`

### 4. 回測功能修復
- 新增 openai==1.54.0 和 httpx==0.27.0 依賴
- 修復 knowledge-admin-api 缺少 openai 模組問題
- 確保版本兼容性（與 rag-orchestrator 一致）

### 5. 前端優化
- 導航文字更新：意圖分類 → 意圖分配
- 恢復意圖分配頁面到知識優化層

## 相關文檔
- docs/CLASSIFICATION_TRACKING_FIX.md
- docs/APPROVAL_FUNCTION_FIX.md
- docs/KNOWLEDGE_IMPORT_SIMILARITY_FIX.md
- APPROVAL_FIX_SUMMARY.md
- KNOWLEDGE_IMPORT_FIX_SUMMARY.md

+++ rag-orchestrator/services/knowledge_import_service.py
@@ -615,7 +615,7 @@ class KnowledgeImportService:
 
                 # 使用資料庫函數檢查是否已存在相似知識
                 result = await conn.fetchrow("""
-                    SELECT * FROM check_knowledge_exists_by_similarity($1::vector, $2)
+                    SELECT * FROM check_knowledge_exists_by_similarity($1::vector, $2::DECIMAL)
                 """, vector_str, threshold)
 
                 if result and (result['exists_in_knowledge_base'] or result['exists_in_review_queue'] or result.get('exists_in_test_scenarios', False)):
